refactored i2c sensors and added accelerometer

+++ I2cDevice.h
@@ -2,11 +2,13 @@
 #define I2CDEVICE
 
 #include "Arduino.h"
+#include <Wire.h>
+#include "Globals.h"
 
 class I2cDevice{
 protected:
 	bool ready = false;
-	uint8_t deviceAddress;
+	int deviceAddress;
 
 	void writeRegister(uint8_t reg,uint8_t val);
 	uint8_t readRegister(uint8_t reg);
